requirements.txt
pydantic==2.6.*
openai==1.14.*
python-dotenv==1.0.*
pyyaml==6.0.*
pytest==8.0.*
requests==2.31.*
requests-mock==1.11.*
streamlit==1.32.*
